requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
requests
beautifulsoup4
python-dotenv
langchain
langchain-google-genai
google-generativeai
pydantic
aiohttp
